updated java version

+++ src/main/java/com/anotationpractice/repository/UserRepository.java
@@ -1,10 +1,11 @@
 package com.anotationpractice.repository;
 
+
 import com.anotationpractice.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
 
 }
